feat: applyed role and authentication authorization

+++ Buy-NET.API/Controllers/CategoryController.cs
@@ -1,12 +1,13 @@
 using Buy_NET.API.Contracts.Category;
 using Buy_NET.API.Services.Interfaces.CategoryServiceInterfaces;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Buy_NET.API.Controllers;
 
 [ApiController]
 [Route("categorias")]
-public class CategoryController : ControllerBase
+public class CategoryController : BaseControllerBuyNet
 {
     private readonly ICategoryService  _categoryService;
 
@@ -16,10 +17,16 @@ public class CategoryController : ControllerBase
     }
 
     [HttpPost]
+    [Authorize]
     public async Task<IActionResult> Create(CategoryRequestContract category)
     {
         try
         {
+            var role = GetLoggedInUserRole();
+            if (role != "Admin")
+            {
+                return Unauthorized("Voce não possui permissão para criar uma categoria");
+            }
             return Created("", await _categoryService.Create(category));
         }
         catch (Exception ex)
@@ -30,6 +37,7 @@ public class CategoryController : ControllerBase
     }
     
     [HttpGet]
+    [Authorize]
     public async Task<IActionResult> Get()
     {
         try
@@ -43,6 +51,7 @@ public class CategoryController : ControllerBase
     }
     
     [HttpGet("search")]
+    [Authorize]
     public async Task<IActionResult> Search([FromQuery] long? id, [FromQuery] string name)
     {
         try
@@ -67,10 +76,16 @@ public class CategoryController : ControllerBase
     }
 
     [HttpPut("{id}")]
+    [Authorize]
     public async Task<IActionResult> Update(long id, CategoryRequestContract category)
     {
         try
         {
+            var role = GetLoggedInUserRole();
+            if (role != "Admin")
+            {
+                return Unauthorized("Voce não possui permissão para atualizar uma categoria");
+            }
             return Ok(await _categoryService.Update(id, category));
         }
         catch (Exception ex)
@@ -80,10 +95,16 @@ public class CategoryController : ControllerBase
     }
     
     [HttpDelete("{id}")]
+    [Authorize]
     public async Task<IActionResult> Delete(long id)
     {
         try
         {
+            var role = GetLoggedInUserRole();
+            if (role != "Admin")
+            {
+                return Unauthorized("Voce não possui permissão para deletar uma categoria");
+            }
             await _categoryService.Delete(id);
             return NoContent();
         }
