+++ Buy-NET.API/Controllers/UserController.cs
@@ -1,13 +1,14 @@
 using System.Security.Authentication;
 using Buy_NET.API.Contracts.User;
 using Buy_NET.API.Services.Interfaces.UserService;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Buy_NET.API.Controllers;
 
 [ApiController]
 [Route("usuarios")]
-public class UserController : ControllerBase
+public class UserController : BaseControllerBuyNet
 {
     private readonly IUserService  _userService;
 
@@ -17,6 +18,7 @@ public class UserController : ControllerBase
     }
 
     [HttpPost]
+    [AllowAnonymous]
     public async Task<IActionResult> Create(UserRequestContract user)
     {
         try
@@ -31,6 +33,7 @@ public class UserController : ControllerBase
     }
 
     [HttpPost("login")]
+    [AllowAnonymous]
     public async Task<IActionResult> Authenticate(UserLoginRequestContract user)
     {
         try
@@ -48,10 +51,16 @@ public class UserController : ControllerBase
     }
     
     [HttpGet]
+    [Authorize]
     public async Task<IActionResult> Get()
     {
         try
         {
+            var role = GetLoggedInUserRole();
+            if (role != "Admin")
+            {
+                return Unauthorized("Voce não possui permissão para buscar todos os usuários");
+            }
             return Ok(await _userService.Get());
         }
         catch (Exception ex)
@@ -61,10 +70,16 @@ public class UserController : ControllerBase
     }
     
     [HttpGet("search")]
+    [Authorize]
     public async Task<IActionResult> Search([FromQuery] long? id, [FromQuery] string email)
     {
         try
         {
+            var role = GetLoggedInUserRole();
+            if (role != "Admin")
+            {
+                return Unauthorized("Voce não possui permissão para buscar um usuário por nome ou id");
+            }
             if (id.HasValue)
             {
                 return Ok(await _userService.GetById(id.Value));
@@ -85,10 +100,17 @@ public class UserController : ControllerBase
     }
 
     [HttpPut("{id}")]
-    public async Task<IActionResult> Update(long id, UserRequestContract user)
+    [Authorize]
+    public async Task<IActionResult> Update(long id, UserUpdateRequestContract user)
     {
         try
         {
+            var userId = GetLoggedInUser();
+            var role = GetLoggedInUserRole();
+            if (userId != id && role != "Admin")
+            {
+                return Unauthorized("Voce não pode atualizar o registro de outro usuário");
+            }
             return Ok(await _userService.Update(id, user));
         }
         catch (Exception ex)
@@ -98,10 +120,16 @@ public class UserController : ControllerBase
     }
     
     [HttpDelete("{id}")]
+    [Authorize]
     public async Task<IActionResult> Delete(long id)
     {
         try
         {
+            var role = GetLoggedInUserRole();
+            if (role != "Admin")
+            {
+                return Unauthorized("Voce não possui permissão para deletar um usuário");
+            }
             await _userService.Delete(id);
             return NoContent();
         }
